+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v1.1.0
 - Feature: Integration of order synchronization in the toolbox webservice
+- Feature: Retrieving the status of an order in the toolbox webservice
 
 # v1.0.2
 - Feature: Outsourcing of the toolbox via webservice